+++ src/main/java/com/ayoshiko/productivebeesgenesis/mixin/ae2/Ae2CentrifugeFactoryMixin.java
@@ -5,6 +5,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import net.minecraft.core.Direction;
 
 import appeng.api.networking.IGridNode;
+import appeng.api.util.AECableType;
 
 import org.jetbrains.annotations.Nullable;
 
@@ -35,5 +36,10 @@ public abstract class Ae2CentrifugeFactoryMixin implements IAe2OutputHost {
 	@Override
 	public @Nullable IGridNode getGridNode(Direction dir) {
 		return IAe2OutputHost.resolveGridNode(this, dir);
+	}
+
+	@Override
+	public AECableType getCableConnectionType(Direction dir) {
+		return IAe2OutputHost.resolveCableConnectionType(dir);
 	}
 }